Updated to include logs
Created update_log_file function: Defined a new function that takes the attack statistics as parameters and updates the log file with the current date/time and the attack statistics. The function retrieves the log file path, formats the log entry, and appends it to the log file.

Called update_log_file function: After calculating the attack statistics in the attack function, the update_log_file function is called to update the log file with the attack statistics.

With these updates, each time an attack is run, the log file named "FoodPanddos Logs.txt" (created if it doesn't exist) will be updated with the date/time of the attack and the corresponding attack statistics. The log file is stored in the same directory as the script file.